requirements.txt
streamlit
pymongo
python-dotenv
requests
